third_party/blink/web_tests/external/wpt/html/browsers/windows/document-domain-removed-iframe-expected.txt

This is a testharness.js-based test.
PASS No removal, no document.domain
PASS Removal, no document.domain
PASS No removal, document.domain
FAIL Removal, document.domain assert_equals: expected (string) "" but got (undefined) undefined
Harness: the test ran to completion.
